refactor: command package 위치 이동(domain -> controller)
