+++ src/main/java/chess/controller/command/Command.java
@@ -1,4 +1,4 @@
-package chess.domain.command;
+package chess.controller.command;
 
 import chess.view.OutputView;
