+++ src/main/java/chess/controller/command/Start.java
@@ -1,4 +1,4 @@
-package chess.domain.command;
+package chess.controller.command;
 
 import chess.domain.ChessGame;
 import chess.view.OutputView;
@@ -13,7 +13,7 @@ public class Start implements Command {
     @Override
     public void execute(String rawInputCommand, final OutputView outputView) {
         chessGame.start();
-        
+
         outputView.printBoard(chessGame.board().getValue());
     }
 }
