+++ src/main/java/chess/controller/command/Status.java
@@ -1,4 +1,4 @@
-package chess.domain.command;
+package chess.controller.command;
 
 import chess.domain.ChessGame;
 import chess.view.OutputView;
